feat(issue-repository): IssueQueryRepository 인터페이스화
- IssueQueryRepository 인터페이스 생성
- JdbcIssueQueryRepository 구현체 생성
- 서비스 계층에서 구현체가 아닌 인터페이스를 참조하도록 변경
